Implement Create Bill & show Bill list

+++ Sources/app/src/main/java/bkhn/et/hospitalbill/ui/login/LoginFragment.java
@@ -74,7 +74,7 @@ public class LoginFragment extends BaseFragment implements ILoginView {
     @Override
     public void openMainScreen(int type) {
         Intent i = new Intent(mContext, MainActivity.class);
-        i.putExtra(AppConstants.Main.EXTRA_JOB_TYPE, type);
+        i.putExtra(AppConstants.Main.EXTRA_JOB_TYPE, (long)type);
         startActivity(i);
         mActivity.finish();
     }
